+++ src/game/nlcore/elements/condition.ts
@@ -123,7 +123,7 @@ export class Condition<Closed extends true | false = false> extends Actionable<n
         }
 
         this.conditions.ElseIf.push({
-            condition: Lambda.isLambda(condition) ? condition : new Lambda(condition),
+            condition: Lambda.from(condition),
             action: this.construct(Array.isArray(action) ? action : [action])
         });
         return this.chain() as Closed extends false ? Proxied<Condition<true>, Chained<LogicAction.Actions>> : never;
@@ -206,7 +206,7 @@ export class Condition<Closed extends true | false = false> extends Actionable<n
     private createIfCondition(
         condition: Lambda | LambdaHandler<boolean>, action: ActionStatements
     ): Proxied<Condition, Chained<LogicAction.Actions>> {
-        this.conditions.If.condition = condition instanceof Lambda ? condition : new Lambda(condition);
+        this.conditions.If.condition = Lambda.from(condition);
         this.conditions.If.action = this.construct(action);
 
         const chained = this.chain();